# docs/swagger.yaml
basePath: /x_ray
definitions:
...
  dto.CreateSalaryReq:
    properties:
      benefit:
        type: integer
      bonus:
        type: integer
      company_id:
        type: string
      experience:
        type: integer
      job_title:
        type: string
      position:
        type: string
      salary_avg:
        description: average net salary
        type: integer
      salary_max:
        description: maximum net salary
        type: integer
      salary_min:
        description: minimum net salary
        type: integer
      user_id:
        type: string
    type: object
...
  dto.SalaryRes:
    properties:
      benefit:
        type: integer
      bonus:
        type: integer
      company_id:
        type: string
      experience:
        type: integer
      id:
        type: string
      job_title:
        type: string
      position:
        type: string
      salary_avg:
        type: integer
      salary_max:
        type: integer
      salary_min:
        type: integer
    type: object
  dto.UserRes:
    properties:
      avatar_url:
        type: string
      created_at:
        type: integer
      email:
        type: string
      google_id:
        type: string
      id:
        type: string
      last_login:
        type: integer
      name:
        type: string
      position:
        type: string
      updated_at:
        type: integer
    type: object
  localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      time:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: this Backend is for x-ray, a company review platform. this backend/project
    was build with the intention to providing transparency to the job seekers, and
    also to give the company a chance to improve their work environment. this project
    is still in early stage, so expect some breaking changes in the future.
  title: X-ray API
  version: "1.0"
paths:
...
  /salary:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of salaries with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Salaries fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SalaryRes'
                  type: array
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Salaries by Pagination
      tags:
      - Salary
    post:
      consumes:
      - application/json
      description: Creates a new salary information entry for a company (requires
        authentication)
      parameters:
      - description: Salary details
        in: body
        name: salary
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSalaryReq'
      produces:
      - application/json
      responses:
        "200":
          description: Salary created successfully
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "401":
          description: Unauthorized - user not authenticated
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create Salary
      tags:
      - Salary
  /salary/{salary_id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific salary entry
      parameters:
      - description: Salary ID (UUID)
        in: path
        name: salary_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Salary fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SalaryRes'
              type: object
        "400":
          description: Missing salary_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: Salary not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Salary by ID
      tags:
      - Salary
  /salary/company/{company_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all salary information for a specific company with pagination
      parameters:
      - description: Company ID (UUID)
        in: path
        name: company_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Salaries fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SalaryRes'
                  type: array
              type: object
        "400":
          description: Missing company_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Salaries by Company ID
      tags:
      - Salary
  /salary/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all salary entries submitted by a specific user with
        pagination
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Salaries fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SalaryRes'
                  type: array
              type: object
        "400":
          description: Missing user_id or invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Salaries by User ID
      tags:
      - Salary
  /user:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of users with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserRes'
                  type: array
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get Users by Pagination
      tags:
      - User
  /user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves user information using their user ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserRes'
              type: object
        "400":
          description: Missing user_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get User by ID
      tags:
      - User
  /user/google/{google_id}:
    get:
      consumes:
      - application/json
      description: Retrieves user information using their Google OAuth ID
      parameters:
      - description: Google OAuth ID
        in: path
        name: google_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserRes'
              type: object
        "400":
          description: Missing google_id parameter
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/localization.StandardResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/localization.StandardResponse'
      summary: Get User by Google ID
      tags:
      - User
swagger: "2.0"